refactor: move ts_class to package's root
