+++ tstore/tsdf/extensions/array.py
@@ -10,8 +10,8 @@ import numpy as np
 import pandas as pd
 from pandas.api.extensions import ExtensionArray
 
+from tstore.ts_class import TS
 from tstore.tsdf.extensions.ts_dtype import TSDtype
-from tstore.tsdf.ts_class import TS
 
 #### Notes
 # https://pandas.pydata.org/pandas-docs/stable/reference/extensions.html
