+++ tstore/tsdf/extensions/ts_dtype.py
@@ -13,7 +13,7 @@ from pandas.api.extensions import (
     register_extension_dtype,
 )
 
-from tstore.tsdf.ts_class import TS
+from tstore.ts_class import TS
 
 #### Notes
 # https://pandas.pydata.org/pandas-docs/stable/reference/extensions.html
